requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
psycopg2-binary==2.9.9
pandas==2.2.0
openpyxl==3.1.5
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
Jinja2==3.1.2